// todo-webapp/src/apiClient/apiClient.ts
export interface UserViewModel {
    id: number;
    fullName: string;
}

export interface UserTaskViewModel {
    id: number;
    name?: string | undefined;
    deadline: string;
    userId?: number | undefined;
    status: string;
    groupId: number;
}

export interface UserTaskDTO {
    name?: string | undefined;
    deadline?: string | undefined;
    userId?: number;
    status?: number;
    groupId: number;
}

export enum TaskStatus {
    New = 0,
    InProgress = 1,
    Completed = 2,
}